PythonBackend/requirements.txt
websockets==12.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1